README.md
Hello! :octocat: :smile:

I just updated the README file for VerseBox. It now gives a more in-depth description of this app and what it's all about, it displays the links to my GitHub, Tumblr, Company Website, and Company Tumblr, and it gives you a call to action! I hope that this README is better than it used to be, and if there are any problems, feel free to contact me or open up an issue.

Thank You and God Bless!

---Jonathan Bruce (@JwBZx)

(:octocat: :+1:) :crown: :-1: