README.md
Hello!

I updated the README file today with some super small changes to the text. I just spaced out the text to make it easier to read, and I removed the line between the title at the top ("VerseBox") and the "By JwBZx" part. So yeah, I just super small changes to the README file, but now it's better than before!

Thank You for reading the README file and God Bless!

---Jonathan Bruce (@JwBZx)

#fuseday #bestnightoftheweek