Created index.js
Hello GitHub! It's been a while since I've coded and especially since I coded for fun, but this afternoon, sitting in a Starbucks, listening to the "Thrice Upon a Time" soundtrack, I was like let's do it!

So I created this simple React Native index.js for the front page of VerseBox, a project that I would like to pick back up!

That's all for now! Maybe one day I'll fix trackr lol

.j.