Hello! Got back into coding a little bit today!
Hello GitHub! I listened to a React Native podcast on the way home from work today and it reminded me about how much fun I had coding and it made me want to try to code again tonight!

Spent time reorganizing the code for fun and then coded a little title screen that needs work.

Need to make this an Expo app and get back in it
Goodnight GitHub!